+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 UNRELEASED
 ----------
 
+* [ [#375](https://github.com/digitalfabrik/lunes-cms/issues/375) ] 375: Add created by filter to CMS
+* [ [#739](https://github.com/digitalfabrik/lunes-cms/issues/739) ] 739: Add option to duplicate jobs
 
 2026.4.7
 --------