+++ CHANGELOG.md
@@ -6,6 +6,7 @@ UNRELEASED
 * [ [#739](https://github.com/digitalfabrik/lunes-cms/issues/739) ] 739: Add option to duplicate jobs
 * [ [#687](https://github.com/digitalfabrik/lunes-cms/issues/687) ] 687: Add automatical audio generation after CSV import
 * [ [#775](https://github.com/digitalfabrik/lunes-cms/issues/775) ] 775: Fix job csv re-import
+* [ [#777](https://github.com/digitalfabrik/lunes-cms/issues/777) ] 777: Add plural and plural article to CSV import
 
 2026.4.7
 --------