+++ CHANGELOG.md
@@ -7,6 +7,7 @@ UNRELEASED
 * [ [#687](https://github.com/digitalfabrik/lunes-cms/issues/687) ] 687: Add automatical audio generation after CSV import
 * [ [#775](https://github.com/digitalfabrik/lunes-cms/issues/775) ] 775: Fix job csv re-import
 * [ [#777](https://github.com/digitalfabrik/lunes-cms/issues/777) ] 777: Add plural and plural article to CSV import
+* [ [#787](https://github.com/digitalfabrik/lunes-cms/issues/787) ] 787: Fix contrast in dark mode
 
 2026.4.7
 --------